feat: Fix upload image count and update submit/upload function
